// lib/game-config/generated.ts
export const cards = [
  {
    "name": "R1",
    "rarity": "R",
    "image": "/images/cards/player/R1/card.webp"
  },
  {
    "name": "R2",
    "rarity": "R",
    "image": "/images/cards/player/R2/card.webp"
  },
  {
    "name": "R3",
    "rarity": "R",
    "image": "/images/cards/player/R3/card.webp"
  },
  {
    "name": "R4",
    "rarity": "R",
    "image": "/images/cards/player/R4/card.webp"
  },
  {
    "name": "SR1",
    "rarity": "SR",
    "image": "/images/cards/player/SR1/card.webp"
  },
  {
    "name": "SR2",
    "rarity": "SR",
    "image": "/images/cards/player/SR2/card.webp"
  },
  {
    "name": "SR3",
    "rarity": "SR",
    "image": "/images/cards/player/SR3/card.webp"
  },
  {
    "name": "SR4",
    "rarity": "SR",
    "image": "/images/cards/player/SR4/card.webp"
  },
  {
    "name": "SSR1",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR1/card.webp"
  },
  {
    "name": "SSR2",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR2/card.webp"
  },
  {
    "name": "SSR3",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR3/card.webp"
  },
  {
    "name": "SSR4",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR4/card.webp"
  },
  {
    "name": "UR1",
    "rarity": "UR",
    "image": "/images/cards/player/UR1/card.webp"
  },
  {
    "name": "UR2",
    "rarity": "UR",
    "image": "/images/cards/player/UR2/card.webp"
  },
  {
    "name": "UR3",
    "rarity": "UR",
    "image": "/images/cards/player/UR3/card.webp"
  }
] as const;
// ...
export const battleResultOdds = [
  {
    "result": "Attack",
    "odds": 2.0
  },
  {
    "result": "Defense",
    "odds": 24.0
  },
  {
    "result": "Coin",
    "odds": 22.0
  },
  {
    "result": "Reply",
    "odds": 8.0
  },
  {
    "result": "Bar",
    "odds": 1000.0
  },
  {
    "result": "SingleChance",
    "odds": 37.0
  },
  {
    "result": "DoubleChance",
    "odds": 114.0
  },
  {
    "result": "TripleChance",
    "odds": 800.0
  }
] as const;

export const visualPatternsByResult = {
  "Attack": [
    {
      "cards": [
        "Coin",
        "Attack",
        "Empty"
      ],
      "weight": 15.0
    },
    {
      "cards": [
        "Defense",
        "Attack",
        "Empty"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Reply",
        "Attack",
        "Empty"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Attack",
        "Reply",
        "Empty"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Attack",
        "Empty",
        "Empty"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Bar",
        "Empty",
        "Attack"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Coin",
        "Coin",
        "Attack"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Defense",
        "Defense",
        "Attack"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Attack",
        "Attack",
        "Attack"
      ],
      "weight": 3.0
    },
    {
      "cards": [
        "Coin",
        "Attack",
        "Attack"
      ],
      "weight": 6.0
    },
    {
      "cards": [
        "Attack",
        "Reply",
        "Attack"
      ],
      "weight": 6.0
    }
  ],
  "Defense": [
    {
      "cards": [
        "Defense",
        "Defense",
        "Defense"
      ],
      "weight": 100.0
    }
  ],
  "SingleChance": [
    {
      "cards": [
        "Chance",
        "Reply",
        "Bar"
      ],
      "weight": 30.0
    },
    {
      "cards": [
        "Bar",
        "Chance",
        "Empty"
      ],
      "weight": 20.0
    },
    {
      "cards": [
        "Coin",
        "Chance",
        "Empty"
      ],
      "weight": 20.0
    },
    {
      "cards": [
        "Bar",
        "Bar",
        "Chance"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Reply",
        "Empty",
        "Chance"
      ],
      "weight": 20.0
    }
  ],
  "DoubleChance": [
    {
      "cards": [
        "Chance",
        "Chance",
        "Bar"
      ],
      "weight": 40.0
    },
    {
      "cards": [
        "Bar",
        "Chance",
        "Chance"
      ],
      "weight": 30.0
    },
    {
      "cards": [
        "Chance",
        "Reply",
        "Chance"
      ],
      "weight": 30.0
    }
  ],
  "TripleChance": [
    {
      "cards": [
        "Chance",
        "Chance",
        "Chance"
      ],
      "weight": 100.0
    }
  ],
  "Coin": [
    {
      "cards": [
        "Coin",
        "Coin",
        "Coin"
      ],
      "weight": 100.0
    }
  ],
  "Reply": [
    {
      "cards": [
        "Reply",
        "Reply",
        "Reply"
      ],
      "weight": 100.0
    }
  ],
  "Bar": [
    {
      "cards": [
        "Bar",
        "Bar",
        "Bar"
      ],
      "weight": 100.0
    }
  ],
  "Empty": [
    {
      "cards": [
        "Coin",
        "Coin",
        "Empty"
      ],
      "weight": 25.0
    },
    {
      "cards": [
        "Defense",
        "Defense",
        "Empty"
      ],
      "weight": 25.0
    },
    {
      "cards": [
        "Reply",
        "Reply",
        "Bar"
      ],
      "weight": 25.0
    },
    {
      "cards": [
        "Bar",
        "Bar",
        "Empty"
      ],
      "weight": 10.0
    },
    {
      "cards": [
        "Reply",
        "Empty",
        "Empty"
      ],
      "weight": 15.0
    }
  ]
} as const;